+++ MU_time.h
@@ -8,7 +8,7 @@ inline unsigned long long musnake::updateTime() {
 	static unsigned flameCount = 0;
 	static unsigned timeCount = 0;
 	unsigned long long tv = SDL_GetTicks();
-	timeDelta = tv - timeVal;
+	timeDelta = (unsigned long)(tv - timeVal);
 	timeVal = tv;
 	flameCount++;
 	timeCount += timeDelta;
@@ -24,7 +24,7 @@ inline unsigned long long musnake::getTimeVal() {
 	return timeVal;
 }
 
-inline unsigned long long musnake::getTimeDelta() {
+inline unsigned long musnake::getTimeDelta() {
 	return timeDelta;
 }
 
@@ -86,7 +86,7 @@ inline void musnake::triggerDelayFunc(DelayFunc** list) {
 	DelayFunc* np;
 	
 	while (np = *list) {
-		if (np->time > getTimeVal()) break;
+		if (np->time > (long long)getTimeVal()) break;
 		*list = np->next;
 		np->func(np->arg);
 		delete np;
